Check Images Exists (Before Looping)
Added check of Image objects before looping to try remove them.

+++ DeleteAll.php
@@ -164,23 +164,25 @@ class MagentoConnection{
 		foreach($this->ProductList as $Product):
 			//Get Images For Product
 			$Product->Images = $this->GetProductImageList($Product->product_id);
-				//Loop Through Images
-				foreach($Product->Images as $Image):
-					$Deletion_FaultCount = 0;
-					Delete_TryAgain:
-					//Delete Images
-					$DeletionResult = $this->DeleteProductImage($Product->product_id, $Image->file);
-					if($DeletionResult != true){
-						$Deletion_FaultCount++;
-						if($this->retry_attempts < $Deletion_FaultCount){
-							goto Delete_TryAgain;
+				if(isset($products->Images)){
+					//Loop Through Images
+					foreach($Product->Images as $Image):
+						$Deletion_FaultCount = 0;
+						Delete_TryAgain:
+						//Delete Images
+						$DeletionResult = $this->DeleteProductImage($Product->product_id, $Image->file);
+						if($DeletionResult != true){
+							$Deletion_FaultCount++;
+							if($this->retry_attempts < $Deletion_FaultCount){
+								goto Delete_TryAgain;
+							}
+						}else{
+							$this->Deletions[] = array($Product->product_id, $Image->file);
+							$LogString = 'Deleted - ProductID:' . $Product->product_id . '. File:' . $Image->file;
+							$this->LogInfo($LogString);
 						}
-					}else{
-						$this->Deletions[] = array($Product->product_id, $Image->file);
-						$LogString = 'Deleted - ProductID:' . $Product->product_id . '. File:' . $Image->file;
-						$this->LogInfo($LogString);
-					}
-				endforeach;
+					endforeach;
+				}
 		endforeach;
 	}
 }
